requirements.txt
# Core
pydantic>=2.0
pydantic-settings>=2.0

# LLM
groq>=0.9.0

# Embeddings
sentence-transformers>=2.7.0
torch>=2.0.0

# NLP
spacy>=3.7.0

# PDF / DOCX parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

# Vector store
chromadb>=0.5.0

# ML / Ranking
lightgbm>=4.3.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0
shap>=0.45.0

# UI
streamlit>=1.35.0

# Config / env
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
